Refatorando etapa 4 de construir cenário; Criando contexto para cenário;
