+++ src/screens/step4/types.ts
@@ -1,23 +1,10 @@
-import { WritableDraft } from "immer/dist/internal";
-import { Scenario } from "types/Plan";
+import { Scenario, Threat } from "types/Plan";
 
 export interface SuggestionList {
   id: string;
   cobrade: string;
   risco: string;
   medida: string;
-}
-
-export interface DuplicateScenariosLines {
-  attr: keyof Scenario;
-  value: any;
-  draftScenariosList: WritableDraft<Scenario>[];
-}
-
-export interface FilterScenarioList {
-  list: any[];
-  attr: keyof Scenario;
-  value: any;
 }
 
 export interface ScenarioDTO {
@@ -34,4 +21,5 @@ export interface ScenarioDTO {
     attr: keyof Scenario,
     value: any,
   ) => boolean;
+  addedCobrades: Array<Threat>;
 }
